pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "cflow-platform"
version = "0.1.6"
description = "CFlow Platform (Phase 1 wrapper) for MCP public API, env verification, and hook installation"
requires-python = ">=3.11"
authors = [{ name = "Cerebral" }]
dependencies = [
    "build>=1.3.0",
    "twine>=6.1.0",
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "chromadb>=0.5.5",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "psutil>=7.0.0",
    "supabase>=2.18.1",
    "psycopg[binary]>=3.2.1",
]
readme = "README.md"
license = { text = "MIT" }
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Operating System :: OS Independent",
]

[tool.hatch.build.targets.wheel]
packages = ["cflow_platform"]

[project.scripts]
cflow-install-hooks = "cflow_platform.install_hooks:install"
cflow-verify-env = "cflow_platform.verify_env:cli"
cflow-test-runner = "cflow_platform.core.test_runner:cli"
cflow-agent-loop = "cflow_platform.core.agent_loop:cli"
cflow-plan-agent = "cflow_platform.core.agents.plan_agent:cli"
cflow-implement-agent = "cflow_platform.core.agents.implement_agent:cli"
cflow-test-agent = "cflow_platform.core.agents.test_agent:cli"
cflow-sync = "cflow_platform.cli.sync_supervisor:cli"
cflow-caef = "cflow_platform.cli.caef_cli:cli"
cflow-setup-cursor = "cflow_platform.cli.setup_cursor:cli"
cflow-one-touch = "cflow_platform.cli.one_touch_installer:cli"
cflow-memory-check = "cflow_platform.cli.memory_check:cli"
cflow-memory-watch = "cflow_platform.cli.memory_watch:cli"
cflow-codebase-vectorize = "cflow_platform.cli.codebase_vectorize:cli"
cflow-migrate-supabase = "cflow_platform.cli.migrate_supabase:cli"
cflow-docs-backfill = "cflow_platform.cli.docs_backfill:cli"
cflow-chroma-reset = "cflow_platform.cli.chroma_reset:cli"
cflow-tasks-backfill = "cflow_platform.cli.tasks_backfill:cli"
cflow-desktop-notify = "cflow_platform.cli.desktop_commander:cli"
cflow-provider-probe = "cflow_platform.cli.provider_probe:cli"
cflow-codegen = "cflow_platform.cli.codegen_cli:cli"
cflow-provider-aws = "cflow_platform.cli.provider_aws:cli"
cflow-chroma-tools = "cflow_platform.cli.chroma_tools:cli"
cflow-supabase-backfill = "cflow_platform.cli.supabase_backfill:cli"



